build: unblock animation loaders on AppleClang

+++ src/JSystem/J3DGraphLoader/J3DAnmLoader.cpp
@@ -7,6 +7,7 @@
 #include "JSystem/J3DGraphLoader/J3DAnmLoader.h"
 #include "JSystem/J3DGraphAnimator/J3DAnimation.h"
 #include "JSystem/JSupport/JSupport.h"
+#include <stdint.h>
 #include <dolphin/os.h>
 #if PLATFORM_PC
 #include "pal/pal_j3d_swap.h"
@@ -393,10 +394,10 @@ void J3DAnmFullLoader_v15::setAnmVtxColor(J3DAnmVtxColorFull* dst,
 
     for (s32 i = 0; i < dst->mAnmTableNum[0]; i++)
         dst->mAnmVtxColorIndexData[0][i].mpData =
-            (void*)((s32)indexPtr0 + (s32)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr0 + (uintptr_t)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
     for (s32 i = 0; i < dst->mAnmTableNum[1]; i++)
         dst->mAnmVtxColorIndexData[1][i].mpData =
-            (void*)((s32)indexPtr1 + (s32)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr1 + (uintptr_t)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
 
     dst->mColorR = JSUConvertOffsetToPtr<u8>(data, data->mRValuesOffset);
     dst->mColorG = JSUConvertOffsetToPtr<u8>(data, data->mGValuesOffset);
@@ -674,10 +675,10 @@ void J3DAnmKeyLoader_v15::setAnmVtxColor(J3DAnmVtxColorKey* dst,
 
     for (s32 i = 0; i < dst->mAnmTableNum[0]; i++)
         dst->mAnmVtxColorIndexData[0][i].mpData =
-            (void*)((s32)indexPtr0 + (s32)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr0 + (uintptr_t)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
     for (s32 i = 0; i < dst->mAnmTableNum[1]; i++)
         dst->mAnmVtxColorIndexData[1][i].mpData =
-            (void*)((s32)indexPtr1 + (s32)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
+            (void*)((uintptr_t)indexPtr1 + (uintptr_t)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
 
     dst->mColorR = JSUConvertOffsetToPtr<s16>(data, data->mRValOffset);
     dst->mColorG = JSUConvertOffsetToPtr<s16>(data, data->mGValOffset);
